Create firsttry
This needs the .json file (in this case 2020-05-04_factbook.json) to be found in the path given on line 11. This path and file name may be changed to mine from a different set of data. All data originate from https://github.com/iancoleman/cia_world_factbook_api#data (Data --> Historical). Code was created and tested in VSCode environment v.1.49.0.